dash/src/requirements.txt
dash
pandas
requests
notebook
jupyter-dash
requests
psycopg2-binary
numpy
